Move tests from `tests` into `src`
