--- app/src/main/utils.ts
@@ -1,4 +1,0 @@
-// Temporary function to have something to test
-export function sum(a: number, b: number): number {
-  return a + b;
-}
