+++ app/src/renderer/features/session/sessionSlice.test.ts
