--- app/tests/main/utils.test.ts
@@ -1,6 +1,0 @@
-import { expect, test } from "vitest";
-import { sum } from "../../src/main/utils";
-
-test("adds 1 + 2 to equal 3", () => {
-  expect(sum(1, 2)).toBe(3);
-});
